Move symlink follow helper func into per-os files
The Windows symlinkDirHelper() func needs to call
filepath.EvalSymlinks() due to issues with calling Stat() on symlinks
(see the following issues for discussion):

https://github.com/golang/go/issues/19870
https://github.com/golang/go/issues/19922
